Storage backend refactor/improvements
Disk and GCS had a lot of shared functionality (blob and index files),
which led to some duplicated code.  Further, they so happened to have
compatible on-disk formats, but this wasn’t enforced actively.

Therefore, now there’s a common packFileBackend that stores chunks in
pack files and maintains index files. It in turn is implemented in
terms of a fileStorage interface, of which there are now disk and GCS
implementations (which are now quite simple).